Introduced TimedEventManager Support for Task Timeout pmSafePQ's and pmThread's priority type is now a template parameter Introduced memSectionSpecifics. Each memsection now has it's own memory maps (inflight, lazy) and has it's own locks CUDA build fix pmSignalWait can now wait

+++ source/code/inc/pmPublicDefinitions.h
@@ -274,7 +274,7 @@ namespace pm
 		unsigned long taskId;		/* Meant for application to assign and identify tasks */
 		unsigned short priority;	/* By default, this is set to max priority level (0) */
         pmSchedulingPolicy policy;  /* By default, this is SLOW_START */
-        unsigned long timeOutInSecs;/* By default, this is max possible value in unsigned long */
+        int timeOutInSecs;          /* By default, this is max possible value in signed int, negative values mean no timeout */
         bool autoFetchOutputMem;    /* By default, this is true */        
 		pmClusterHandle cluster;	/* Unused */
